Remove complex callback logic.
This commit restores the performance optimization from 7ad5848, which
was accidentally broken in e5485f6e -- the next commit!

It goes back to a simpler way to factor out the `search` logic between
the three use cases:

1. find the solution to a problem when there's exactly one
2. find all solutions to a problem
3. determine if there's more than one solution to a problem

and to abort the search as soon as a solution is found in cases 2 and 3.

Previously, the callback was in charge of two aspects of the logic
(recording solutions and determining whether to abort the search).

Now these responsibilities are split between two arguments, and that's
easier to understand.